requirements.txt
requests
amadeus
ddgs
beautifulsoup4
google-adk
python-dotenv
